feat: 콘솔앱용 dao 구현
