--- src/main/java/nextstep/ReservationDAO.java
@@ -1,94 +1,0 @@
-package nextstep;
-
-import roomescape.domain.Reservation;
-import roomescape.domain.Theme;
-
-import java.sql.*;
-
-public class ReservationDAO {
-
-    private Connection getConnection() {
-        Connection con = null;
-        try {
-            con = DriverManager.getConnection("jdbc:h2:~/test;AUTO_SERVER=true", "sa", "");
-            System.out.println("정상적으로 연결되었습니다.");
-        } catch (SQLException e) {
-            System.err.println("연결 오류:" + e.getMessage());
-            e.printStackTrace();
-        }
-        return con;
-    }
-
-    private void close(Connection con) {
-        try {
-            if (con != null)
-                con.close();
-        } catch (SQLException e) {
-            System.err.println("con 오류:" + e.getMessage());
-        }
-    }
-
-    public void addReservation(Reservation reservation) {
-        Connection con = getConnection();
-        try {
-            String sql = "INSERT INTO reservation (date, time, name, theme_name, theme_desc, theme_price) VALUES (?, ?, ?, ?, ?, ?);";
-            PreparedStatement ps = con.prepareStatement(sql, new String[]{"id"});
-            ps.setDate(1, Date.valueOf(reservation.getDate()));
-            ps.setTime(2, Time.valueOf(reservation.getTime()));
-            ps.setString(3, reservation.getName());
-            ps.setString(4, reservation.getTheme().getName());
-            ps.setString(5, reservation.getTheme().getDesc());
-            ps.setInt(6, reservation.getTheme().getPrice());
-            ps.executeUpdate();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-        close(con);
-    }
-
-    public Reservation findReservation(Long id) {
-        Reservation reservation = null;
-        Connection con = getConnection();
-        try {
-            String sql = "SELECT * FROM reservation WHERE id = ?;";
-            PreparedStatement ps = con.prepareStatement(sql);
-            ps.setLong(1, id);
-            ResultSet rs = ps.executeQuery();
-            if (rs.next()) {
-                reservation = new Reservation(
-                        rs.getLong("id"),
-                        rs.getDate("date").toLocalDate(),
-                        rs.getTime("time").toLocalTime(),
-                        rs.getString("name"),
-                        new Theme(
-                                rs.getString("theme_name"),
-                                rs.getString("theme_desc"),
-                                rs.getInt("theme_price")
-                        )
-                );
-            } else {
-                throw new IllegalArgumentException("데이터 없음 비상비상");
-            }
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-        close(con);
-        return reservation;
-    }
-
-    public int removeReservation(Long id) {
-        int count = 0;
-        Connection con = getConnection();
-        try {
-            String sql = "DELETE FROM reservation WHERE id = ?;";
-            PreparedStatement ps = con.prepareStatement(sql);
-            ps.setLong(1, id);
-            count = ps.executeUpdate();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-        close(con);
-        return count;
-    }
-
-}
